requirements.txt
flask
python-dotenv
google-generativeai
langgraph
langchain-core
